requirements.txt
python-dotenv==1.2.2
polars>=0.20.0
pymysql>=1.1.0
sqlalchemy>=1.4.36,<2.0.0
apache-airflow-providers-mysql
